[+] useGroupsDrawer hook
